move TEI specific test in to TEI test suite
